+++ backend/src/main/java/org/example/database/entity/HeartRate.java
@@ -22,7 +22,7 @@ public class HeartRate {
     private double averageBpm;
 
     @Column(name = "timestamp_utc")
-    private int timestampUtc;
+    private Long timestampUtc;
 
     @Column(name = "timestamp_ms")
     private int timestampMs;
@@ -56,11 +56,11 @@ public class HeartRate {
         this.averageBpm = averageBpm;
     }
 
-    public int getTimestampUtc() {
+    public Long getTimestampUtc() {
         return timestampUtc;
     }
 
-    public void setTimestampUtc(int timestampUtc) {
+    public void setTimestampUtc(Long timestampUtc) {
         this.timestampUtc = timestampUtc;
     }
