+++ backend/src/main/java/org/example/database/service/GnssService.java
@@ -42,7 +42,7 @@ public class GnssService {
         return gnssRepository.findById(id);
     }
 
-    public List<Gnss> findGnssByTimestampUtcBetween(int start, int end) {
+    public List<Gnss> findGnssByTimestampUtcBetween(Long start, Long end) {
         return gnssRepository.findByTimestampUtcBetween(start, end);
     }
